Convert from references, not values
This solves the lifetime/borrow/move issues when constructing errors,
since errors now just containg references. The only gain from converting
values was avoiding cloning the vector of vars for Exists. It really
should work this way anyway: the main function will try to convert ULC
to TRC and TRC to SQL, and if anything fails, it should be able to print
the entire original value in addition to the specific context.